+++ app/system/modules/theme/app/views/theme.js
@@ -1,0 +1,7 @@
+import UIkit from 'uikit';
+import Icons from '../../../../../../node_modules/uikit/dist/js/uikit-icons';
+
+// loads the Icon plugin
+UIkit.use(Icons);
+
+import '../../assets/less/theme.less';